Browser console warnings fix
